requirements.txt
sentence-transformers
spacy
nltk
pdfplumber
PyPDF2
python-docx
pandas
dateparser
fastapi
uvicorn
httpx
python-dotenv
pydantic
torch
numpy
psycopg[binary]
